+++ src/Microsoft.Diagnostics.Runtime/DacImplementation/DacRuntime.cs
@@ -15,7 +15,7 @@ using Microsoft.Diagnostics.Runtime.Utilities;
 
 namespace Microsoft.Diagnostics.Runtime.DacImplementation
 {
-    internal sealed unsafe class DacRuntime : IAbstractRuntime, IAbstractThreadProvider, IAbstractThreadPoolProvider
+    internal sealed unsafe class DacRuntime : IAbstractRuntime, IAbstractThreadProvider
     {
         private readonly IDataReader _dataReader;
         private readonly ThreadStoreData _threadStore;
@@ -354,53 +354,6 @@ namespace Microsoft.Diagnostics.Runtime.DacImplementation
         }
 
         public string? GetJitHelperFunctionName(ulong address) => _sos.GetJitHelperFunctionName(address);
-
-        ////////////////////////////////////////////////////////////////////////////////
-        // ThreadPool
-        ////////////////////////////////////////////////////////////////////////////////
-        public IAbstractThreadPoolProvider? LegacyThreadPoolHelpers => this;
-
-        public bool GetLegacyThreadPoolData(out LegacyThreadPoolInfo result)
-        {
-            HResult hr = _sos.GetThreadPoolData(out ThreadPoolData data);
-            result = new()
-            {
-                AsyncTimerCallbackCompletionFPtr = data.AsyncTimerCallbackCompletionFPtr,
-                CpuUtilization = data.CpuUtilization,
-                CurrentLimitTotalCPThreads = data.CurrentLimitTotalCPThreads,
-                FirstUnmanagedWorkRequest = data.FirstUnmanagedWorkRequest,
-                HillClimbingLog = data.HillClimbingLog,
-                HillClimbingLogFirstIndex = data.HillClimbingLogFirstIndex,
-                HillClimbingLogSize = data.HillClimbingLogSize,
-                MaxFreeCPThreads = data.MaxFreeCPThreads,
-                MaxLimitTotalCPThreads = data.MaxLimitTotalCPThreads,
-                MaxLimitTotalWorkerThreads = data.MaxLimitTotalWorkerThreads,
-                MinLimitTotalCPThreads = data.MinLimitTotalCPThreads,
-                MinLimitTotalWorkerThreads = data.MinLimitTotalWorkerThreads,
-                NumCPThreads = data.NumCPThreads,
-                NumFreeCPThreads = data.NumFreeCPThreads,
-                NumIdleWorkerThreads = data.NumIdleWorkerThreads,
-                NumRetiredCPThreads = data.NumRetiredCPThreads,
-                NumRetiredWorkerThreads = data.NumRetiredWorkerThreads,
-                NumTimers = data.NumTimers,
-                NumWorkingWorkerThreads = data.NumWorkingWorkerThreads,
-            };
-
-            return hr;
-        }
-
-        public bool GetLegacyWorkRequestData(ulong workRequest, out LegacyWorkRequestInfo workRequestInfo)
-        {
-            bool res = _sos.GetWorkRequestData(workRequest, out WorkRequestData workRequestData);
-            workRequestInfo = new()
-            {
-                Function = workRequestData.Function,
-                Context = workRequestData.Context,
-                NextWorkRequest = workRequestData.NextWorkRequest,
-            };
-
-            return res;
-        }
 
         ////////////////////////////////////////////////////////////////////////////////
         // COM
